Se implemento validacion al delete de bahia

+++ src/Views/BayView.java
@@ -100,10 +100,21 @@ public class BayView extends View{
             input.printCorrecto("Bahia encontrada: ");
             render.printTable(bay);
 
-            this.controller.delete(code);
+            String confirmDelete = "";
+            while (!confirmDelete.equalsIgnoreCase("S") && !confirmDelete.equalsIgnoreCase("N")) {
+                confirmDelete = input.getString("¿Estás seguro de querer eliminar esta bahia? (S/N): ");
 
+                if (confirmDelete.equalsIgnoreCase("S")) {
+                    this.controller.delete(code);
+                    input.printCorrecto("Bahia eliminada correctamente");
+                } else if (confirmDelete.equalsIgnoreCase("N")) {
+                    input.printAlerta("Eliminación cancelada");
+                } else {
+                    input.printAlerta("Por favor solo ingresar S/N");
+                }
+            }
         } catch (Exception e) {
-            input.printAlerta("Bahia no encontrado");
+            input.printAlerta("Bahia no encontrada");
         }
     }
